Require Ruby 2.6.x
When the Gemfile said "ruby RUBY_VERSION" and I did a `bundle install` with
Ruby 2.5 active, the install succeeded, and rewrote the "RUBY VERSION" section
of Gemfile.lock to use my 2.5 version.

With this restatement of the ruby version requirement, `bundle install` with
Ruby 2.5 active causes an error. This better reflects the README assertion that
Ruby 2.6 is required.